backend/requirements.txt
django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
